initial react-client implementation
